updated importing files section
