+++ workshop_3/4_importing_files.py
@@ -1,0 +1,8 @@
+from src.maths import maths
+from src.music import songs
+
+
+total = maths.add_two_numbers(1, 2)
+print(total)
+
+songs.sing_happy_birthday("Vitali Klitschko")
