--- workshop_3/4_importing_files_different_folder.py
@@ -1,5 +1,0 @@
-import src.maths as maths
-
-total = maths.add_two_numbers(1, 2)
-
-print(total)
