--- workshop_3/5_importing_files_same_folder.py
@@ -1,3 +1,0 @@
-import songs
-
-songs.sing_happy_birthday("Vitali Klitschko")
