--- workshop_3/songs.py
@@ -1,5 +1,0 @@
-def sing_happy_birthday(name):
-    print("Happy birthday to you")
-    print("Happy birthday to you")
-    print(f"Happy birthday dear {name}")
-    print("Happy birthday to you")
