--- workshop_3/src/maths.py
@@ -1,5 +1,0 @@
-def add_two_numbers(first_number, second_number):
-
-    sum_of_numbers = first_number + second_number
-
-    return sum_of_numbers
