+++ workshop_3/src/maths/maths.py
@@ -1,0 +1,20 @@
+def add_two_numbers(first_number, second_number):
+    """
+    Adds two numbers together and returns the result.
+
+    Parameters
+    ----------
+
+    first_number (int) : the first number
+    second_number (int) : the second number
+
+    Returns
+    -------
+
+    int : the sum of the two numbers
+
+    """
+    
+    sum_of_numbers = first_number + second_number
+
+    return sum_of_numbers
