+++ workshop_3/src/music/songs.py
@@ -1,0 +1,19 @@
+def sing_happy_birthday(name):
+    """
+    Prints the lyrics to the Happy Birthday song for a given name
+
+    Parameters
+    ----------
+
+    name (string) : The name of the person whose birthday it is
+
+    Returns
+    -------
+
+    None
+
+    """
+    print("Happy birthday to you")
+    print("Happy birthday to you")
+    print(f"Happy birthday dear {name}")
+    print("Happy birthday to you")
